+++ CHANGES.md
@@ -18,6 +18,7 @@
 - Move `GeojsonSource` to new file `lib/Map/Vector/ProtomapsGeojsonSource.ts`.
 - support URL parameters in a GetLegendGraphic request for a layer without a style configured
 - Enhanced error processing for obtaining user location
+- Replace clipboardjs with native browser clipboard
 - [The next improvement]
 
 #### 8.7.11 - 2024-12-18